Add options to load markdown files for privacy/terms
