+++ next.config.js
@@ -7,4 +7,11 @@ module.exports = {
     includePaths: [path.join(__dirname, "styles")],
   },
   reactStrictMode: true,
+  webpack: (config) => {
+    config.module.rules.push({
+      test: /\.md$/,
+      use: "raw-loader",
+    });
+    return config;
+  },
 };
