+++ package.json
@@ -21,14 +21,15 @@
   "dependencies": {
     "formik": "^2.2.6",
     "json2md": "^1.10.0",
+    "markdown-to-jsx": "^7.1.1",
     "next": "10.0.8",
     "next-i18next": "^8.1.1",
     "notifications-node-client": "^5.0.2",
     "pino": "^6.11.1",
+    "raw-loader": "^4.0.2",
     "react": "^17.0.1",
     "react-dom": "17.0.1",
-    "uuid": "^8.3.2",
-    "markdown-to-jsx": "^7.1.1"
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@babel/core": "^7.13.10",
