+++ public/static/content/en/terms.md
@@ -1,0 +1,1 @@
+# Terms
