+++ public/static/content/fr/privacy.md
@@ -1,0 +1,1 @@
+# Prive
