+++ public/static/content/fr/terms.md
@@ -1,0 +1,1 @@
+# Avis
